pyproject.toml
[tool.poetry]
name = "starbelly"
version = "2.0.0-dev"
description = "Streaming crawler with a graphical user interface"
authors = ["Mark E. Haase <mehaase@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.6.2"
aiohttp-socks = "^0.3.9"
beautifulsoup4 = "^4.9.1"
cchardet = "^2.1.6"
feedparser = "^5.2.1"
formasaurus = {url = "https://hyperiongray.s3-us-west-2.amazonaws.com/formasaurus-0.9.0.tar.gz"}
lxml = "^4.5.1"
protobuf = "^3.12.2"
psutil = "^5.7.0"
python-dateutil = "^2.8.1"
python-mimeparse = "^1.6.0"
pytz = "==2020.1"
rethinkdb = {url = "https://hyperiongray.s3-us-west-2.amazonaws.com/rethinkdb-2.4.7%2Bstarbelly.tar.gz"}
robotexclusionrulesparser = "^1.7.1"
trio = "^0.15.1"
trio-asyncio = "^0.11.0"
trio-websocket = "^0.8.0"
w3lib = "1.22.0"
watchdog = "^0.10.2"
yarl = "^1.4.2"
scikit-learn = ">=0.18" # required for formasaurus
scipy = "^1.4.1" # required for formasaurus
sklearn-crfsuite = ">=0.3.1" # required for formasaurus
joblib = "^1.2.0" # required for formasaurus
tldextract = "^2.2.2" # required for formasaurus
soft404 = "^0.2.1" # for detecting soft 404 pages

[tool.poetry.dev-dependencies]
coverage = "^5.1"
coveralls = "^2.0.0"
pytest-cov = "^2.9.0"
pytest-mock = "^3.1.0"
sphinxcontrib-trio = "^1.1.2"
sphinx-rtd-theme = "^0.4.3"
pytest-trio = "^0.6.0"
trustme = "^0.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
